Fix a bunch of ODR related CodeChecker issues
CodeChecker was complaining about a bunch of ODR issues in the test code
because the tests were defined in header files. Get rid of those
complaints by defining the tests for each module in a cpp file and
having a separate test executable per module.

This is better anyway because it reduces coupling between the test
modules.
