+++ CMakeLists.txt
@@ -8,7 +8,10 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 add_subdirectory(3rdparty)
+
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+enable_testing()
+include(GoogleTest)
 
 project(Sq
     VERSION 0.0.1
