--- src/parser/test/test_parser.cpp
@@ -1,1 +1,0 @@
-#include "test/test_parser.h"
