+++ src/serialization/test/CMakeLists.txt
@@ -1,13 +1,8 @@
-set(SQ_ST_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/include")
-set(SQ_ST_HEADERS_DIR "${SQ_ST_INCLUDE_DIR}/test")
-set(SQ_ST_SRC_DIR "${CMAKE_CURRENT_SOURCE_DIR}")
-
-add_library(sq_serialization_test
-    "${SQ_ST_HEADERS_DIR}/test_serialization.h"
-    "${SQ_ST_SRC_DIR}/test_serialization.cpp"
+add_executable(sq-serialization-test
+    "${CMAKE_CURRENT_SOURCE_DIR}/test_serialization.cpp"
 )
-target_include_directories(sq_serialization_test PUBLIC "${SQ_ST_INCLUDE_DIR}")
-set_target_properties(sq_serialization_test PROPERTIES CXX_CLANG_TIDY "")
-target_link_libraries(sq_serialization_test PUBLIC sq_serialization)
-target_link_libraries(sq_serialization_test PUBLIC sq_results_test)
-target_link_libraries(sq_serialization_test PUBLIC gtest)
+set_target_properties(sq-serialization-test PROPERTIES CXX_CLANG_TIDY "")
+target_link_libraries(sq-serialization-test sq_serialization)
+target_link_libraries(sq-serialization-test sq_results_test_util)
+target_link_libraries(sq-serialization-test gtest_main)
+gtest_discover_tests(sq-serialization-test)
