--- src/system/test/test_system.cpp
@@ -1,1 +1,0 @@
-#include "test/test_system.h"
