--- src/test.cpp
@@ -1,7 +1,0 @@
-#include "test/test_ast.h"
-#include "test/test_common_types.h"
-#include "test/test_parser.h"
-#include "test/test_results.h"
-#include "test/test_serialization.h"
-#include "test/test_system.h"
-#include "test/test_util.h"
